chore(note): move files out of internal directory
