+++ src/components/note/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from "./__internal__/note";
-export * from "./__internal__/note";
+export { default } from './note';
+export * from './note';
