+++ src/components/note/note.d.ts
