+++ src/components/note/note.stories.mdx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import Note from './note.component';
 import {
   ActionPopover,
   ActionPopoverDivider,
   ActionPopoverItem
-} from '../../action-popover';
+} from '../action-popover';
 import { EditorState, ContentState, convertFromHTML } from 'draft-js';
 
 <Meta title="Design System/Note" />
